pyproject.toml
[project]
name = "receiptaggregator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ollama>=0.5.1",
    "lxml>=6.0.0",
    "lxml-html-clean>=0.4.2",
    "pydantic>=2.11.7",
    "google-genai>=1.24.0",
    "scikit-learn>=1.7.0",
    "matplotlib>=3.10.3",
    "numpy>=2.3.1",
    "polars>=1.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "SIM", # flake8-simplify
    "ANN", # flake8-annotations
    "D",   # flake8-docstrings

]

ignore = [
    "E501",
    "B905",
    "D213",
    "D203",
    "D205",
    "D100",
    "D104"
]
exclude = [
    "src/receiptaggregator/string_similarity.py"
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.12.2",
]

